container rebuilds when Gemfile updates
This addresses a perceived issue that when a book updates a gem
dependency, then the bookdocker container would stop working because
bundle exec would error out when it saw that it didn't have a dependent
gem installed.

This fixes the issue by ALWAYS running the docker build command.  In
order to make sure that the bundle build step is cached, we only add the
Gemfile files to the container at build time.

Signed-off-by: Tom Dunlap <tdunlap@pivotal.io>